change to copy just before write

+++ interp.py
@@ -13,10 +13,12 @@ class Pointer:
     privilege: str  # none, read, write
     __match_args__ = ("address", "privilege")
 
+# Increment the reference count of an address
 def increment(addr, mem):
     (v, count) = mem[addr]
     mem[addr] = (v, count + 1)
 
+# Decrement the reference count of an address
 def decrement(addr, mem):
     (v, count) = mem[addr]
     new_count = count - 1
@@ -29,9 +31,9 @@ def share(ptr, mem):
     match ptr:
       case Pointer(addr, priv):
         if priv == 'none':
-            raise Exception('cannot share if have no priveleges')
+            raise Exception('cannot share if have no privileges')
         increment(addr, mem)
-        ptr.privelege = 'read'
+        ptr.privilege = 'read'
         return Pointer(addr, 'read')
       case _:
         raise Exception('expected pointer, not ' + repr(ptr))
@@ -40,9 +42,9 @@ def release(ptr, mem):
     match ptr:
       case Pointer(addr, priv):
         if priv != 'write':
-            raise Exception('cannot release if do not have write privelege')
+            raise Exception('cannot release if do not have write privilege')
         ptr.addr = -1
-        ptr.privelege = 'none'
+        ptr.privilege = 'none'
         return Pointer(addr, 'write')
       case _:
         raise Exception('expected pointer, not ' + repr(ptr))
@@ -53,17 +55,32 @@ def kill(val, mem):
         pass
       case Pointer(addr, priv):
         if priv != 'write':
-            raise Exception('cannot kill if do not have write privelege')
+            raise Exception('cannot kill if do not have write privilege')
         decrement(addr, mem)
         val.addr = -1
-        val.privelege = 'none'
-    
+        val.privilege = 'none'
+
+def bind(val, mem):
+  match val:
+    case Integer(value):
+      return Integer(value)
+    case Pointer(address, priv):
+      if mem[address][1] > 0 and priv != 'read':
+        raise Exception('in bind, cannot steal from owner')
+      increment(address, mem)
+      return val
+    case _:
+      raise Exception('in bind, unhandled value ' + repr(val))
+        
 def copy(val, mem):
     match val:
       case Integer(value):
         return Integer(value)
       case Pointer(address, priv):
-        return share(val, mem)
+        if priv != 'read':
+            raise Exception('in copy, pointer must have read privilege, not ' + priv)
+        increment(address, mem)
+        return Pointer(address, priv)
       case _:
         raise Exception('in copy, unhandled value ' + repr(val))
     
@@ -71,8 +88,8 @@ def read(ptr, mem):
     match ptr:
       case Pointer(addr, priv):
         if priv == 'none':
-            raise Exception('cannot read if do not have read privelege')
-        return copy(mem[addr][0], mem)
+            raise Exception('cannot read if do not have read privilege')
+        return mem[addr][0]
       case _:
         raise Exception('in read expected a pointer, not ' + repr(ptr))
 
@@ -80,10 +97,12 @@ def write(ptr, val, mem):
     match ptr:
       case Pointer(addr, priv):
         if priv != 'write':
-            raise Exception('cannot write if do not have write privelege')
+            raise Exception('cannot write if do not have write privilege')
         (old_val, count) = mem[addr]
+        if count != 1:
+            raise Exception('write requires unique ownership')
         kill(old_val, mem)
-        mem[addr] = (val, count)
+        mem[addr] = (copy(val, mem), count)
       case _:
         raise Exception('in read expected a pointer, not ' + repr(ptr))
     
@@ -97,21 +116,27 @@ def interp_exp(e, env, mem):
       case Prim('add', args):
         left = interp_exp(args[0], env, mem)
         right = interp_exp(args[1], env, mem)
-        return left.value + right.value
+        return Integer(left.value + right.value)
       case Prim('sub', args):
         left = interp_exp(args[0], env, mem)
         right = interp_exp(args[1], env, mem)
-        return left.value - right.value
+        return Integer(left.value - right.value)
       case Prim('neg', args):
-        return - interp_exp(args[0], env, mem).value
+        return Integer(- interp_exp(args[0], env, mem).value)
       case New(arg):
         val = interp_exp(arg, env, mem)
         addr = len(mem)
-        mem[addr] = (val, 1)
+        mem[addr] = (copy(val, mem), 0)
         return Pointer(addr, 'write')
       case Deref(arg):
         ptr = interp_exp(arg, env, mem)
         return read(ptr, mem)
+      case Share(arg):
+        ptr = interp_exp(arg, env, mem)
+        return share(ptr, mem)
+      case Release(arg):
+        ptr = interp_exp(arg, env, mem)
+        return release(ptr, mem)
       case _:
         raise Exception('error in interp_exp, unhandled: ' + repr(e)) 
     
@@ -121,14 +146,15 @@ def interp_stmt(s, env, mem):
       case Return(e):
         return interp_exp(e, env, mem)
       case Init(var, init):
-        env[var] = interp_exp(init, env, mem)
+        val = interp_exp(init, env, mem)
+        env[var] = bind(val, mem)
       case Assign(var, rhs):
         kill(env[var], mem)
         env[var] = interp_exp(rhs, env, mem)
       case Write(lhs, rhs):
         ptr = interp_exp(lhs, env, mem)
         val = interp_exp(rhs, env, mem)
-        write(ptr, val, mem)
+        write(ptr, copy(val, mem), mem)
       case _:
         raise Exception('error in interp_stmt, unhandled: ' + repr(s))
